requirements.txt
adjustText==1.3.0
anndata==0.11.4
anthropic==0.50.0
bbknn==1.6.0
beautifulsoup4==4.13.4
bin2cell==0.3.3
celltypist==1.6.3
dask==2024.9.1
glmpca==0.1.0
holoviews==1.20.2
kneed==0.8.5
loguru==0.7.3
loompy==3.0.8
matplotlib==3.10.1
memory_profiler==0.61.0
MulticoreTSNE==0.1
numpy==2.2.5
pandas==2.2.3
panel==1.6.3
Pillow==9.5.0
Pillow==11.2.1
psutil==5.9.7
psutil==5.9.8
pyarrow==12.0.0
pyensembl==2.3.13
python-dotenv==1.1.0
scanpy==1.11.1
scib==1.1.7
scikit_learn==1.2.2
scipy==1.15.2
seaborn==0.13.2
seagal==2.5.2
spatialdata_io==0.2.0
squidpy==1.6.5
stlearn==0.4.12
tensorflow==2.16.2
tensorflow_macos==2.12.0
torch==2.6.0
tqdm==4.67.1
umap_learn==0.5.7
